# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Lunch Meet",
    "description": "Description"
  },
  "host": "3.35.54.135",
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "로그인 이후 토큰을 여기다가 넣으세요~! Bearer 빠트리지 말구요 !"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/comment/{lunchid}": {
      "get": {
        "tags": ["Comment"],
        "description": "해당 약속 리플 가져오기",
        "parameters": [
          {
            "name": "lunchid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Comment"],
        "description": "해당 약속 리플달기",
        "parameters": [
          {
            "name": "lunchid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "security": [{ "token": [] }],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/comment/{commentid}": {
      "delete": {
        "tags": ["Comment"],
        "description": "해당 약속 코맨트 삭제",
        "security": [{ "token": [] }],
        "parameters": [
          {
            "name": "commentid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/applicant/{lunchid}": {
      "get": {
        "tags": ["Applicant"],
        "description": "해당 점약 신청자 가져오기",
        "parameters": [
          {
            "name": "lunchid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": ["Applicant"],
        "description": "해당 점약 신청",
        "security": [{ "token": [] }],
        "parameters": [
          {
            "name": "lunchid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Applicant"],
        "description": "해당 점약 취소",
        "security": [{ "token": [] }],
        "parameters": [
          {
            "name": "lunchid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}